server/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
cryptography==35.0.0
Django==3.0.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.27.0
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-secret-manager==2.7.0
google-cloud-storage==1.38.0
google-cloud-logging==2.0.2
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
grpcio==1.40.0
gunicorn==20.0.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.0
oauth2==1.9.0.post1
oauthlib==3.1.1
opencv-python==4.5.3.56
openpyxl==3.0.7
packaging==20.9
pandas==1.3.0
pandas-datareader==0.9.0
Pillow==8.2.0
pipenv==2021.5.29
py-geohash-any==1.1
PyJWT==2.1.0
pylint==2.9.6
pytz==2021.1
PyMySQL==1.0.2
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
SQLAlchemy==1.4.25
tweepy==3.10.0
urllib3==1.26.4
virtualenv==20.4.7
virtualenv-clone==0.5.4
whitenoise==5.2.0
sklearn==0.0
